Correct principle to principle as needed for #5
In some cases, this draft is discussing principles for using identities.

In other cases, this draft is using the defined term in this draft for
a Principal and its relationship to an Identifier.

These changes correct the latter where writing it as the former. This
likely seems to be incorrect.